Extension of Viewport widget to support pyramidal StaticStreams. The information of the field of view and meters per pixel is getting to the Stream classes. This information is used to calculate which tiles will be read from the TIFF files (or other types of tiles that implement tiling). So far, there's only one Stream for all the viewports. So changes in the field of view and meters per pixel of one view port is still affecting all the others viewports.
Extension of Viewport widget to support
pyramidal StaticStreams.
The information of the field of view and meters per pixel is getting to the Stream classes. This information is used to calculate which tiles will be read from the TIFF files (or other types of tiles that implement tiling).
So far, there's only one Stream for all the viewports. So changes in the field of view and meters per pixel of one view port is still affecting all the others viewports.